move checks on family to method instead of internal function
Adding the two.stage argument introduced a bug where if 'family'
was not a character string, fitNullModel would fail.
One of the family checks requires the two.stage argument, which
is not passed to the internal .fitNullModel function. Move these
checks to the data.frame method instead (in the form of another
internal function). The internal function now assumes that
'family' is in the form of a function all (e.g., 'gaussian()'),
so modify unit tests accordingly.